refactor: use src folder
