+++ package.json
@@ -3,8 +3,7 @@
   "description": "Simple cross Node.js inteface for using brotli compression",
   "homepage": "https://nicedoc.io/Kikobeats/compress-brotli",
   "version": "1.3.0",
-  "main": "index.js",
-  "types": "index.d.ts",
+  "main": "src/index.js",
   "author": {
     "email": "josefrancisco.verdu@gmail.com",
     "name": "Kiko Beats",
@@ -59,9 +58,7 @@
     "node": ">= 12"
   },
   "files": [
-    "index.js",
-    "index.d.ts",
-    "lib"
+    "src"
   ],
   "scripts": {
     "clean": "rm -rf node_modules",
@@ -98,5 +95,6 @@
   "simple-git-hooks": {
     "commit-msg": "npx commitlint --edit",
     "pre-commit": "npx lint-staged"
-  }
+  },
+  "types": "src/index.d.ts"
 }
