+++ src/merge-options.js
@@ -1,0 +1,18 @@
+'use strict'
+
+module.exports = (defaultOptions = {}, options = {}) => {
+  const params = {
+    ...(defaultOptions.params || {}),
+    ...(options.params || {})
+  }
+
+  return {
+    ...defaultOptions,
+    ...options,
+    ...(Object.keys(params).length
+      ? {
+          params
+        }
+      : {})
+  }
+}
